# AArch64: feature detection on FreeBSD, faster native CRC32 instructions