- assume NEON (ASIMD) support on AArch64, since it's mandatory for all processors
- detect features on FreeBSD >= 12 via `ID_AA64ISAR0_EL1`
- add support for native CRC32 instructions for even more performance (results from Marvell Armada 8040 @ 2ghz)

```
cebda3bb        "/boot/kernel/kernel"   14 ms   975 MB/s    # fallback
cebda3bb        "/boot/kernel/kernel"   2 ms    5735 MB/s   # pmull
cebda3bb        "/boot/kernel/kernel"   1 ms    14049 MB/s  # crc32 instr
```

(not that it affects actual compression by much, gzipping the same file is ~204MB/s with fallback, ~247MB/s with pmull, ~249MB/s with crc32 instructions)