@ebiggers 2019-07-03
Oh, I missed that the benchmark results were in the pull request description.  Can you move them to the commit message instead so they don't get lost?  Thanks!!